Rendering penguin at 200 pixels of distance in an aleatory angle

+++ include/Face.h
@@ -6,11 +6,11 @@
 
 class Face : public GameObject{
 	public:
-		Face(float x, float y);
+		Face(double x, double y);
 		~Face();
 
 		void Damage(int damage);
-		void Update(float dt);
+		void Update(double dt);
 		void Render();
 
 		bool IsDead();
